+++ apps/router-e2e/__e2e__/link-preview/app/(tabs)/home/_layout.tsx
@@ -1,5 +1,10 @@
 import { Stack } from 'expo-router';
 
 export default function Layout() {
-  return <Stack screenOptions={{ headerShown: false, animation: 'slide_from_bottom' }} />;
+  return (
+    <Stack screenOptions={{ headerShown: false, animation: 'slide_from_bottom' }}>
+      <Stack.Screen name="index" />
+      <Stack.Screen name="modal" options={{ presentation: 'modal', headerShown: true }} />
+    </Stack>
+  );
 }